test: add local folder e2e tests
