+++ tests/resources/subdirectories/folder2/metadata.json
@@ -1,0 +1,1 @@
+{"color":  "red", "author":  "team search", "price":  10}
