+++ tests/resources/subdirectories/folder3/test.txt
@@ -1,0 +1,1 @@
+hello
