+++ tests/resources/subdirectories/folder4/test.txt
@@ -1,0 +1,1 @@
+hello
