+++ tests/resources/subdirectories/folder5/test.txt
@@ -1,0 +1,1 @@
+hello
