+++ tests/resources/subdirectories/folder6/metadata.json
@@ -1,0 +1,1 @@
+{"color":  "red", "author":  "team search", "price":  10}
