+++ tests/resources/subdirectories/folder6/test.txt
@@ -1,0 +1,1 @@
+hello
