+++ tests/resources/subdirectories/folder7/test.txt
@@ -1,0 +1,1 @@
+hello
