+++ tests/resources/subdirectories/folder8/test.txt
@@ -1,0 +1,1 @@
+hello
