+++ tests/resources/subdirectories/folder9/test.txt
@@ -1,0 +1,1 @@
+hello
